chore: add test properties
- to be able to deploy on port 80, but test on port 8080
- add ActiveProfiles("test")
- organize packages
- format
- remove unused variables and properties from tests
