+++ src/main/resources/application-test.properties
@@ -1,5 +1,6 @@
 spring.application.name=spring-ecommerce-order
-server.port=80
+
+server.port=8080
 
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
