--- src/test/kotlin/ecommerce/AcceptanceTest.kt
@@ -1,37 +1,0 @@
-package ecommerce
-
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.http.HttpHeaders
-import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.options
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers.print
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.header
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-
-@SpringBootTest
-@AutoConfigureMockMvc
-class AcceptanceTest {
-    @Autowired
-    private lateinit var mockMvc: MockMvc
-
-    @Test
-    fun cors() {
-        mockMvc.perform(
-            options("/api/products")
-                .header(HttpHeaders.ORIGIN, "your host ex)localhost:port")
-                .header(HttpHeaders.ACCESS_CONTROL_REQUEST_METHOD, "GET"),
-        )
-            .andExpect(status().isOk)
-            .andExpect(header().string(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN, "*"))
-            .andExpect(header().string(HttpHeaders.ACCESS_CONTROL_ALLOW_METHODS, ALLOWED_METHOD_NAMES))
-            .andExpect(header().string(HttpHeaders.ACCESS_CONTROL_EXPOSE_HEADERS, HttpHeaders.LOCATION))
-            .andDo(print())
-    }
-
-    companion object {
-        private const val ALLOWED_METHOD_NAMES = "GET,POST,HEAD"
-    }
-}
