+++ src/test/kotlin/ecommerce/ApplicationTests.kt
@@ -1,0 +1,13 @@
+package ecommerce
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+
+@SpringBootTest
+@ActiveProfiles("test")
+class ApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
