+++ src/test/kotlin/ecommerce/config/AcceptanceTest.kt
@@ -1,0 +1,40 @@
+package ecommerce.config
+
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.HttpHeaders
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers
+
+@SpringBootTest
+@ActiveProfiles("test")
+@AutoConfigureMockMvc
+class AcceptanceTest {
+    @Autowired
+    private lateinit var mockMvc: MockMvc
+
+    @Test
+    fun cors() {
+        mockMvc.perform(
+            MockMvcRequestBuilders.options("/api/products")
+                .header(HttpHeaders.ORIGIN, "your host ex)localhost:port")
+                .header(HttpHeaders.ACCESS_CONTROL_REQUEST_METHOD, "GET"),
+        )
+            .andExpect(MockMvcResultMatchers.status().isOk)
+            .andExpect(MockMvcResultMatchers.header().string(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN, "*"))
+            .andExpect(
+                MockMvcResultMatchers.header().string(HttpHeaders.ACCESS_CONTROL_ALLOW_METHODS, ALLOWED_METHOD_NAMES),
+            )
+            .andExpect(MockMvcResultMatchers.header().string(HttpHeaders.ACCESS_CONTROL_EXPOSE_HEADERS, HttpHeaders.LOCATION))
+            .andDo(MockMvcResultHandlers.print())
+    }
+
+    companion object {
+        private const val ALLOWED_METHOD_NAMES = "GET,POST,HEAD"
+    }
+}
