+++ src/test/kotlin/ecommerce/dto/ProductExceptionTest.kt
@@ -6,7 +6,9 @@ import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import org.springframework.test.context.ActiveProfiles
 
+@ActiveProfiles("test")
 class ProductExceptionTest {
     private lateinit var validator: Validator
