+++ src/test/kotlin/ecommerce/model/CartEntityTest.kt
@@ -5,7 +5,9 @@ import ecommerce.repository.CartRepository
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
+import org.springframework.test.context.ActiveProfiles
 
+@ActiveProfiles("test")
 @DataJpaTest
 class CartEntityTest {
     @Autowired
