+++ src/test/kotlin/ecommerce/service/CartServiceTest.kt
@@ -17,7 +17,9 @@ import org.junit.jupiter.api.assertDoesNotThrow
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.context.annotation.Import
+import org.springframework.test.context.ActiveProfiles
 
+@ActiveProfiles("test")
 @DataJpaTest
 @Import(CartService::class)
 class CartServiceTest(
